+++ .github/Changelog.md
@@ -10,7 +10,7 @@
 - Move balances to header
 - Simplified Settings
 - Moved settings to floating action button
+- Implemented Datepicker (for Budget validity by month)
 - ToDo:
-    - Fix Date picker for budget validity
     - Check budget validity calculating remaining budget for the current month in the header
     - Show save button on modification/check/deletion of balances, budgets and open transactions